pyproject.toml
[project]
name = "acme"
version = "0.0.1"
description = "python template"
readme = ".github/README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    "absl-py>=2.1.0",
    "clu>=0.0.12",
    "datasets>=3.6.0",
    "einops>=0.8.1",
    "flax>=0.12.0",
    "huggingface-hub>=0.33.0",
    "jaxtyping>=0.3.2",
    "ml-collections>=1.1.0",
    "optax>=0.2.3",
    "orbax-checkpoint>=0.6.4",
    "safetensors>=0.5.3",
    "pillow>=10.0.0",
    "grain>=0.2.0,<0.3",
]

[dependency-groups]
dev = [
    "pytest>=6.0.0",
    "ruff>=0.12.0",
]
wandb = [
    "wandb>=0.20.1",
]

[tool.ruff]
line-length = 90

[tool.ruff.lint]
select = ["E", "F", "I001"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
	"ignore::UserWarning"
]
markers = [
    "large: larger tests recommended to be run on CI"
]